% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ew_partition_0_1}
\alias{ew_partition_0_1}
\title{Partition [0,1] into S equally sized consecutive segments. Each segment
is given by its midpoint.}
\usage{
ew_partition_0_1(S = 1000)
}
\arguments{
\item{S}{An integer >= 1. The number of segments to partition [0,1] into.}
}
\value{
The vector of segment midpoints,
c(\code{(1-0.5)/S}, \code{(2-0.5)/S}, ... (\code{(S-0.5)/S}).
}
\description{
Partition [0,1] into S equally sized consecutive segments. Each segment
is given by its midpoint.
}
