% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoestimate.R
\name{SSO_graph_plot}
\alias{SSO_graph_plot}
\title{Show the results of a call to SSO_estimate() or SSO_estimate_continous() in a plot.}
\usage{
SSO_graph_plot(SSO_out, cert = 0.95, visual = TRUE)
}
\arguments{
\item{SSO_out}{The output of a call to SSO_estimate() or SSO_estimate_continous().}

\item{cert}{The certainty level for the min and max values of the error fraction.}

\item{visual}{If TRUE, a plot is returned.
If FALSE, a list of specific values is returned.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   If visual is TRUE, a ggplot is returned.
           If visual is FALSE, a named list is returned consisting of
           - the most probable, min, and max values of the error fraction in SSO;
           - and ditto for the client.
}\if{html}{\out{</div>}}
}
\description{
Show the results of a call to SSO_estimate() or SSO_estimate_continous() in a plot.
}
\examples{
SSO_out <- SSO_estimate(k_SSO = 1, n_SSO = 3000, n_client = 500, S_p_SSO = 1000)
SSO_graph_plot(SSO_out)
}
