% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcv_new.R
\name{xcv_new}
\alias{xcv_new}
\title{Split a string by a given character}
\usage{
xcv_new(x, pattern, n = Inf)
}
\arguments{
\item{x}{A character vector}

\item{pattern}{A regular expression, as used in the stringr package. The default is ",".
The pattern is used to describe where to split x.}

\item{n}{This determines the maximum length of each element of the output.}
}
\value{
A character vector
}
\description{
Split a string by a given character
}
\examples{
xcv_new("a,b,c", ",")
}
