% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoestimate.R,
%   R/ssoestimate_before_translation.R
\name{SSO_estimate}
\alias{SSO_estimate}
\title{Estimation of error fraction in SSO and client.}
\usage{
SSO_estimate(k_SSO, n_SSO, n_client, S_p_SSO = 1000, S_k_client = 1000)

SSO_estimate(k_SSO, n_SSO, n_client, S_p_SSO = 1000, S_k_client = 1000)
}
\arguments{
\item{k_SSO}{The number of errors found in Steek_SSO. k_SSO >= 0.}

\item{n_SSO}{The number of posten in Steek_SSO. n_SSO > 0.}

\item{n_client}{The number of posten in Steek_client. n_client > 0.}

\item{S_p_SSO}{The number of segments in the rooster for p. S_p_SSO > 0.}

\item{S_k_client}{The number of segments in the rooster for k_client. S_k_client > 0.}
}
\value{
A list of named elements containing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    - P_SSO, the probability graph of the error fraction in SSO
    - P_client, the probability graph of the error fraction in Steek_client
    - the input parameters
}\if{html}{\out{</div>}}

A list of named elements containing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    - P_SSO, the probability graph of the error fraction in SSO
    - P_client, the probability graph of the error fraction in Steek_client
    - the input parameters
}\if{html}{\out{</div>}}
}
\description{
Estimation of error fraction in SSO and client.

Estimation of error fraction in SSO and client.
}
\examples{
y <- SSO_estimate(
  k_SSO = 1, n_SSO = 3000, n_client = 500,
  S_p_SSO = 1000, S_k_client = 501
)

y <- SSO_estimate(
  k_SSO = 1, n_SSO = 3000, n_client = 500,
  S_p_SSO = 1000, S_k_client = 501
)

}
